+++ angular/buildComponents.sh
@@ -1,4 +1,5 @@
 
+yarn
 cd ./components
 yarn
 ng build channel-value
@@ -9,3 +10,7 @@ cd ..
 # move core folder
 rm -rf ./core
 cp -r ./components/core ./core
+
+# copy index
+rm -rf ./core/index.ts
+cp -r index.ts ./core/index.ts