+++ angular/index.ts
@@ -1,0 +1,5 @@
+import * as emptyState from './core/empty-state/public_api';
+import * as channelValue from './core/channel-value/public_api';
+import * as hero from './core/hero/public_api';
+
+export { channelValue, emptyState, hero };