+++ react/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## v1.1.0
+Adds a new component for EmptyState
+
 ## v1.0.0 
 Fixes a bug in icon size for inline ChannelValue components